requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2026.1.4
charset-normalizer==3.4.4
django==4.2.27
djangorestframework==3.15.2
idna==3.11
numpy==1.24.4
requests==2.32.4
sqlparse==0.5.5
typing-extensions==4.13.2
urllib3==2.2.3
